+++ src/app/cart/components/cart-item/cart-item.component.spec.ts
@@ -3,21 +3,21 @@ import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { By } from "@angular/platform-browser";
 import { DebugElement } from "@angular/core";
 
-import { CartListComponent } from "./cart-list.component";
+import { CartItemComponent } from "./cart-item.component";
 
-describe("CartListComponent", () => {
-    let component: CartListComponent;
-    let fixture: ComponentFixture<CartListComponent>;
+describe("CartItemComponent", () => {
+    let component: CartItemComponent;
+    let fixture: ComponentFixture<CartItemComponent>;
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
-            declarations: [CartListComponent]
+            declarations: [CartItemComponent]
         })
             .compileComponents();
     }));
 
     beforeEach(() => {
-        fixture = TestBed.createComponent(CartListComponent);
+        fixture = TestBed.createComponent(CartItemComponent);
         component = fixture.componentInstance;
         fixture.detectChanges();
     });