+++ src/app/cart/models/cart-item.model.ts
@@ -1,0 +1,7 @@
+import { ProductModel } from "src/app/products/models/product.model";
+
+export interface CartItemModel {
+    product: ProductModel;
+    quantity: number;
+    cost: number;
+}