+++ src/app/cart/services/cart.service.spec.ts
@@ -1,16 +1,16 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async, inject } from '@angular/core/testing';
-import { CartService } from './cart.service';
+import { TestBed, async, inject } from "@angular/core/testing";
+import { CartService } from "./cart.service";
 
-describe('Service: Cart', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [CartService]
+describe("Service: Cart", () => {
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [CartService]
+        });
     });
-  });
 
-  it('should ...', inject([CartService], (service: CartService) => {
-    expect(service).toBeTruthy();
-  }));
+    it("should ...", inject([CartService], (service: CartService) => {
+        expect(service).toBeTruthy();
+    }));
 });