+++ src/app/cart/services/cart.service.ts
@@ -1,30 +1,85 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Optional, signal } from "@angular/core";
 import { ProductModel } from "src/app/products/models/product.model";
-import { ProductsService } from "src/app/products/services/products.service";
+import { MathHelper } from "src/app/shared/math.helper";
+import { CartItemModel } from "../models/cart-item.model";
+import { BehaviorSubject, Observable } from "rxjs";
+import { LocalStorageService } from "src/app/core/services/local-storage.service";
 
 @Injectable({
-    providedIn: 'root'
+    providedIn: "root"
 })
 export class CartService {
-    totalQuantity: number = 0;
-    private cart: Map<ProductModel, number> = new Map();
+    private readonly _cartProducts = new BehaviorSubject<ReadonlyArray<CartItemModel>>([]);
+    private readonly _totalQuantity = signal(0);
+    private readonly _totalCost = signal(0);
+    private readonly _isEmptyCart = signal(true);
 
-    constructor(private readonly productsService:ProductsService) {      
+    totalQuantity = this._totalQuantity.asReadonly();
+    totalCost = this._totalCost.asReadonly();
+    isEmptyCart = this._isEmptyCart.asReadonly();
+
+    constructor(@Optional() private localStorageService: LocalStorageService) { }
+
+    getProducts(): Observable<ReadonlyArray<CartItemModel>> {
+        return this._cartProducts.asObservable();
     }
 
-    addToCart(product?: ProductModel) {
-        const productToAdd = product || this.productsService.getRandomProduct();
+    addProduct(product: ProductModel, quantity: number = 1): void {
+        if (quantity <= 0) {
+            return;
+        }
 
-        this.cart.set(productToAdd, (this.cart.get(productToAdd) || 0) + 1);
-        this.totalQuantity++;
+        const itemToAdd = this._cartProducts.getValue().find(x => x.product === product);
+        if (!itemToAdd) {
+            const cost = MathHelper.round(quantity * product.price);
+            this._cartProducts.next(this._cartProducts.getValue().concat({ product: product, quantity: quantity, cost: cost }));
+            this.updateState(quantity, cost);
+            this.localStorageService?.setItem(product.name, `${new Date(Date.now())}`);
+        } else {
+            this.increaseQuantity(product, quantity);
+        }
     }
 
-    getCartItems(): [ProductModel, number][] {
-        return Array.from(this.cart);
+    removeProduct(product: ProductModel): void {
+        const itemToRemove = this._cartProducts.getValue().find(x => x.product === product);
+        if (itemToRemove) {
+            this._cartProducts.next(this._cartProducts.getValue().filter(x => x.product !== product));
+            this.updateState(-itemToRemove.quantity, -itemToRemove.cost);
+
+            this.localStorageService?.removeItem(product.name);
+        }
     }
 
-    clearCart() {
-        this.cart.clear();
-        this.totalQuantity = 0;
+    removeAllProducts(): void {
+        this._cartProducts.next([]);
+        this.updateState(-this.totalQuantity(), -this.totalCost());
+
+        this.localStorageService?.clear();
+    }
+
+    increaseQuantity(product: ProductModel, quantity: number): void {
+        this.changeQuantity(product, quantity);
+    }
+
+    decreaseQuantity(product: ProductModel, quantity: number): void {
+        this.changeQuantity(product, -quantity);
+    }
+
+    private changeQuantity(product: ProductModel, diffQuantity: number): void {
+        this._cartProducts.next(this._cartProducts.getValue().map(item =>
+            item.product === product
+                ? {
+                    product: product,
+                    quantity: item.quantity + diffQuantity,
+                    cost: MathHelper.round((item.quantity + diffQuantity) * product.price)
+                } : item
+        ));
+        this.updateState(diffQuantity, diffQuantity * product.price);
+    }
+
+    private updateState(quantityToAdd: number, costToAdd: number): void {
+        this._totalQuantity.update(value => value + quantityToAdd);
+        this._totalCost.update(value => MathHelper.round(value + costToAdd));
+        this._isEmptyCart.set(!this._cartProducts.getValue().length);
     }
 }