+++ src/app/core/services/gen-id.generator.ts
@@ -1,0 +1,6 @@
+export function* genID() {
+    let id = 1;
+    while (id < Number.MAX_SAFE_INTEGER) {
+        yield id++;
+    }
+}