+++ src/app/core/services/generated-string.token.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from "@angular/core";
+
+export const GeneratedStringToken = new InjectionToken<(n: number) => string>("GeneratedString");