+++ src/app/products/services/products.service.ts
@@ -1,16 +1,19 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable } from "@angular/core";
 import { ProductModel } from "../models/product.model";
 import { Category } from "../enums/category.enum";
 import { GeneratorService } from "src/app/shared/services/generator.service";
+import { MathHelper } from "src/app/shared/math.helper";
+import { GeneratedStringToken } from "src/app/core/services/generated-string.token";
 
 @Injectable({
-    providedIn: 'root'
+    providedIn: "root"
 })
 export class ProductsService {
 
-    private readonly products: ProductModel[]
+    private readonly products: ProductModel[];
 
-    constructor(private readonly generatorService: GeneratorService) {
+    constructor(private readonly generatorService: GeneratorService,
+        @Inject(GeneratedStringToken) private generatedString: (n: number) => string) {
         this.products = this.generatorService.generateArray(100, (i) => this.generateRandomProduct(i));
     }
 
@@ -22,27 +25,14 @@ export class ProductsService {
         return this.products[this.generatorService.random(this.products.length)];
     }
 
-    private generateRandomProduct(index: number): ProductModel {
+    private generateRandomProduct(_: number): ProductModel {
         const categoryValues = Object.values(Category);
         return {
-            name: `Product_${++index}`,
-            description: this.generateRandomString(Math.floor(Math.random() * 100)),
-            price: parseFloat((Math.random() * 1000).toFixed(2)),
+            name: `Product[id:${this.generatorService.getNewID()}]`,
+            description: this.generatedString(Math.floor(Math.random() * 100)),
+            price: MathHelper.round(Math.random() * 100),
             category: categoryValues[this.generatorService.random(categoryValues.length)] as Category,
             isAvailable: Math.random() > 0.3
         };
-    }
-
-
-    private generateRandomString(length: number): string {
-        const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        let randomString = '';
-
-        for (let i = 0; i < length; i++) {
-            const randomIndex = this.generatorService.random(characters.length);
-            randomString += characters.charAt(randomIndex);
-        }
-
-        return randomString;
     }
 }