+++ src/app/shared/services/generator.service.spec.ts
@@ -1,16 +1,16 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async, inject } from '@angular/core/testing';
-import { GeneratorService } from './generator.service';
+import { TestBed, async, inject } from "@angular/core/testing";
+import { GeneratorService } from "./generator.service";
 
-describe('Service: Generator', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [GeneratorService]
+describe("Service: Generator", () => {
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [GeneratorService]
+        });
     });
-  });
 
-  it('should ...', inject([GeneratorService], (service: GeneratorService) => {
-    expect(service).toBeTruthy();
-  }));
+    it("should ...", inject([GeneratorService], (service: GeneratorService) => {
+        expect(service).toBeTruthy();
+    }));
 });